refactor: 메소드 기능에 맞게 클래스 간 메소드 이동
