+++ src/main/java/racingcar/RacingCarController.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
@@ -14,7 +15,20 @@ public class RacingCarController {
         this.racingCars = racingCars;
     }
 
-    public List<RacingCar> getRacingCars() {
+    public RacingCarController(int amount) {
+        this(RacingCarController.createRacingCars(amount));
+    }
+
+    /**
+     * 주어진 수 만큼 자동차를 만들어 리스트에 담아 리턴합니다.
+     * @param amount
+     * @return
+     */
+    public static List<RacingCar> createRacingCars(int amount) {
+        List<RacingCar> racingCars = new ArrayList<>(amount);
+        for (int i = 0; i < amount; i++) {
+            racingCars.add(new RacingCar());
+        }
         return racingCars;
     }
 
@@ -50,5 +64,17 @@ public class RacingCarController {
      */
     boolean isMovable(int value) {
         return value >= 4;
+    }
+    /**
+     * 자동차의 현재 위치를 받아서 리스트에 담습니다.
+     * @param racingCarController
+     * @return
+     */
+    public List<Integer> getNowDistance() {
+        List<Integer> nowDistances = new ArrayList<>();
+        for (RacingCar racingCar : this.racingCars) {
+            nowDistances.add(racingCar.getDistance());
+        }
+        return nowDistances;
     }
 }
